Added max step
